client_max_body_size logic done; search for server and location now in Server.cpp and no more in <method>_request.cpp; intern error request page problem location_name not removed now resolved

+++ includes/LocationConfig.hpp
@@ -20,7 +20,7 @@ class LocationConfig
 		std::string							get_root();
 		std::string							get_path();
 		std::string							get_index();
-		int									get_client_max_body_size();
+		size_t								get_client_max_body_size();
 		bool								get_autoindex(bool autoindex);
 		std::map<std::string, std::string>&	get_map_location();
